extract buildFilters and chooseMeal out of fancy-menu
