+++ source/views/menus/components/fancy-menu.js
@@ -1,4 +1,5 @@
 // @flow
+
 import React from 'react'
 import {View, StyleSheet} from 'react-native'
 import {connect} from 'react-redux'
@@ -11,23 +12,18 @@ import type {
   ProcessedMealType,
   MenuItemContainerType,
 } from '../types'
-import flatten from 'lodash/flatten'
-import filter from 'lodash/filter'
-import find from 'lodash/find'
 import fromPairs from 'lodash/fromPairs'
-import map from 'lodash/map'
 import size from 'lodash/size'
 import values from 'lodash/values'
-import uniq from 'lodash/uniq'
 import {ListSeparator, ListSectionHeader} from '../../components/list'
-import type {FilterType, PickerType} from '../../components/filter'
+import type {FilterType} from '../../components/filter'
 import {applyFiltersToItem} from '../../components/filter'
-import {fastGetTrimmedText} from '../../../lib/html'
 import SimpleListView from '../../components/listview'
 import {NoticeView} from '../../components/notice'
-import {findMeal} from '../lib/find-menu'
 import {FilterMenuToolbar} from './filter-menu-toolbar'
 import {FoodItemRow} from './food-item-row'
+import {chooseMeal} from '../lib/choose-meal'
+import {buildFilters} from '../lib/build-filters'
 
 type FancyMenuPropsType = TopLevelViewPropsType & {
   applyFilters: (filters: FilterType[], item: MenuItemType) => boolean,
@@ -48,7 +44,6 @@ const styles = StyleSheet.create({
   },
 })
 
-
 class FancyMenuView extends React.Component {
   static defaultProps = {
     applyFilters: applyFiltersToItem,
@@ -63,93 +58,7 @@ class FancyMenuView extends React.Component {
     }
 
     const foodItemsArray = values(foodItems)
-    filters = this.buildFilters(foodItemsArray, menuCorIcons, meals, now)
-    this.props.onFiltersChange(filters)
-  }
-
-  buildFilters(
-    foodItems: MenuItemType[],
-    corIcons: MasterCorIconMapType,
-    meals: ProcessedMealType[],
-    now: momentT
-  ): FilterType[] {
-    // Format the items for the stations filter
-    const stations = flatten(meals.map(meal => meal.stations))
-    const stationLabels = uniq(stations.map(station => station.label))
-    const allStations = stationLabels.map(label => ({title: label}))
-
-    // Grab the labels of the COR icons
-    const allDietaryRestrictions = map(corIcons, cor => ({
-      title: cor.label,
-      image: cor.image ? {uri: cor.image} : null,
-      detail: cor.description ? fastGetTrimmedText(cor.description) : '',
-    }))
-
-    // Decide which meal will be selected by default
-    const mealOptions = meals.map(m => ({label: m.label}))
-    const selectedMeal = this.findMeal(meals, [], now)
-
-    // Check if there is at least one special in order to show the specials-only filter
-    const stationNames = selectedMeal.stations.map(s => s.label)
-    const shouldShowSpecials = filter(foodItems, item =>
-      item.special && stationNames.includes(item.station)).length >= 1
-
-    return [
-      {
-        type: 'toggle',
-        key: 'specials',
-        enabled: shouldShowSpecials,
-        spec: {
-          label: 'Only Show Specials',
-          caption: 'Allows you to either see only the "specials" for today, or everything the location has to offer (e.g., condiments.)',
-        },
-        apply: {
-          key: 'special',
-        },
-      },
-      {
-        type: 'picker',
-        key: 'meals',
-        enabled: true,
-        spec: {
-          title: "Today's Menus",
-          options: mealOptions,
-          selected: {label: selectedMeal.label},
-        },
-        apply: {
-          key: 'label',
-        },
-      },
-      {
-        type: 'list',
-        key: 'stations',
-        enabled: false,
-        spec: {
-          title: 'Stations',
-          options: allStations,
-          mode: 'OR',
-          selected: allStations,
-        },
-        apply: {
-          key: 'station',
-        },
-      },
-      {
-        type: 'list',
-        key: 'dietary-restrictions',
-        enabled: false,
-        spec: {
-          title: 'Dietary Restrictions',
-          showImages: true,
-          options: allDietaryRestrictions,
-          mode: 'AND',
-          selected: [],
-        },
-        apply: {
-          key: 'cor_icon',
-        },
-      },
-    ]
+    this.props.onFiltersChange(buildFilters(foodItemsArray, menuCorIcons, meals, now))
   }
 
   props: FancyMenuPropsType;
@@ -168,27 +77,9 @@ class FancyMenuView extends React.Component {
     })
   }
 
-  findMeal(meals: ProcessedMealType[], filters: FilterType[], now: momentT): ProcessedMealType {
-    const mealChooserFilter: ?PickerType = find(filters, f => f.type === 'picker' && f.spec.title === "Today's Menus")
-    let selectedMeal = meals[0]
-
-    if (mealChooserFilter && mealChooserFilter.spec.selected) {
-      let label = mealChooserFilter.spec.selected.label
-      selectedMeal = meals.find(meal => meal.label === label)
-    } else {
-      selectedMeal = findMeal(meals, now)
-    }
-
-    if (!selectedMeal) {
-      selectedMeal = {label: '', stations: [], starttime: '0:00', endtime: '0:00'}
-    }
-
-    return selectedMeal
-  }
-
   renderSectionHeader = (sectionData: MenuItemType[], sectionName: string) => {
     const {filters, now, meals} = this.props
-    const {stations} = this.findMeal(meals, filters, now)
+    const {stations} = chooseMeal(meals, filters, now)
     const menu = stations.find(m => m.label === sectionName)
     const note = menu ? menu.note : ''
 
@@ -220,7 +111,7 @@ class FancyMenuView extends React.Component {
       meals,
     } = this.props
 
-    const {label: mealName, stations: stationMenus} = this.findMeal(meals, filters, now)
+    const {label: mealName, stations: stationMenus} = chooseMeal(meals, filters, now)
 
     const filteredByMenu = stationMenus
       .map(menu => [
